fix bug syntax

+++ ios/RNTwilioVoice/RNTwilioVoice.m
@@ -265,7 +265,7 @@ RCT_REMAP_METHOD(getActiveCall,
     
     if([userInfo objectForKey:@"type"]){
         if([[userInfo objectForKey:@"type"] isEqualToString:@"connect_expert_request"]){
-            NSInteger answerNowId = [userInfo objectForKey:@"answer_now_id"];
+            int answerNowId = [userInfo objectForKey:@"answer_now_id"];
             NSDictionary *data = @{ answer_now_id : answerNowId};
             [self sendEventWithName:@"requestConnectExpert" body:data];
         }
